--- client/src/components/join/Join.css
@@ -1,7 +1,0 @@
-input[type="submit"]:disabled {
-  opacity: 60%;
-}
-
-input[type="submit"]:disabled:hover {
-  color: var(--primary);
-}
