+++ client/src/components/joinModal/index.ts
@@ -1,0 +1,3 @@
+import Join from "./JoinModal";
+
+export default Join;
