+++ client/src/components/makeModal/index.ts
@@ -1,0 +1,3 @@
+import HomeModal from "./MakeModal";
+
+export default HomeModal;
